Starting fixing issues with tests from project refactoring
