+++ lib/factories/post_factory.rb
@@ -5,7 +5,7 @@ require_relative '../models/post'
 module Factories
   ##
   # This class is a factory for parsing post text and creating a correseponding post model
-  class PostFactory
+  class PostFactory < BaseFactory
     LEAD = '{: .lead}'
     BREAK = '<!–-break-–>'
